refactor: 认证架构调整
