+++ sea-auth/pom.xml
@@ -23,7 +23,7 @@
         <dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-webflux</artifactId>
-	 </dependency>
+	    </dependency>
 
         <dependency>
             <groupId>com.alibaba.cloud</groupId>
@@ -38,12 +38,6 @@
         <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-lang3</artifactId>
-        </dependency>
-
-         <dependency>
-            <groupId>io.jsonwebtoken</groupId>
-            <artifactId>jjwt</artifactId>
-            <version>0.12.6</version> <!-- 或使用较新版本 -->
         </dependency>
 
         <dependency>
