+++ sea-auth/src/main/java/com/example/sea/auth/SeaAuthApplication.java
@@ -11,10 +11,10 @@ import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
  */
 @EnableDiscoveryClient
 @SpringBootApplication(scanBasePackages = {"com.example.sea.auth", "com.example.sea.common"} )
-public class AuthApplication {
+public class SeaAuthApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(AuthApplication.class, args);
+		SpringApplication.run(SeaAuthApplication.class, args);
 	}
 
 }
