--- sea-common/sea-common-security/src/main/java/com/example/config/ResourceServerConfig.java
@@ -1,26 +1,0 @@
-package com.example.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-public class ResourceServerConfig {
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-          .authorizeHttpRequests(auth -> auth
-            .requestMatchers("/api/public").permitAll()
-            .anyRequest().hasAuthority("SCOPE_read")
-          )
-          .csrf(csrf -> csrf.disable())
-          .oauth2ResourceServer(oauth2 -> oauth2
-            .jwt(Customizer.withDefaults())
-          );
-        return http.build();
-    }
-
-}
