+++ sea-gateway/src/main/java/com/example/sea/gateway/SeaGatewayApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
 @EnableDiscoveryClient
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"com.example.sea.gateway", "com.example.sea.common"} )
 public class SeaGatewayApplication {
 
     public static void main(String[] args) {
